fix executor choking on some expressions
For example: (5 / 1 2) / 5

The parser sees this as Vector(5 / Vector(1 2)), which the executor
thought is a matrix in a matrix. This is correct, but not necessarily
why the limit was set up in the first place.

This example should just lead to it being accepted, which it now does:

Matrix(Matrix(a)) => Matrix(a)