parser: reallow negataions in ints
So that -1 is parsed is an unit -1 instead of neg(1), which should
improve speed during runtime.

This was originally removed in 4cc46e1131ed569b7bc69d3fc8f2883650d84636,
since it bound `5-5` to `5 -5`, but this was actually an error in that
vectors just needed space between every item. So this restriction is not
necessary any more.